requirements.txt
# --- core api ---
fastapi==0.115.9
uvicorn==0.34.2
python-multipart
pydantic

# --- ml & processing ---
tensorflow-cpu
scikit-learn==1.6.1
pandas==2.0.3
numpy
scipy==1.11.1
joblib==1.3.1
h5py
protobuf

# --- outils (sans verrous bloquants) ---
nltk
pillow
matplotlib
seaborn
tqdm
requests
python-dotenv